requirements.txt
pydantic>=2.5.0
Pillow>=10.0.0
